Car spawner 1.2
