--- Assets/Cars.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 738dbd9101891c94b99864862d85721a
-folderAsset: yes
-timeCreated: 1517947500
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
