--- Assets/Cars/CarScript.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 695675e8d691c1f43a2ef36d61563489
-timeCreated: 1517934355
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
