--- Assets/Cars/CoolCar.prefab.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 82ca7c422514696418f02b1e065e32e0
-timeCreated: 1517944666
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
