--- Assets/Cars/Wheels.mat.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: d4495427034271f49ab0b151ec60b3d5
-timeCreated: 1517945909
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
