+++ Assets/Prefab/Car/CarModel.prefab
@@ -232,7 +232,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!1 &1777184251417642
 GameObject:
   m_ObjectHideFlags: 0
