+++ Assets/Prefab/Car/CarSpawner.prefab
@@ -52,14 +52,14 @@ Transform:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1194571183833680}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 40, y: 2.5, z: 5}
+  m_LocalRotation: {x: 0, y: -1, z: 0, w: 0}
+  m_LocalPosition: {x: 78.79511, y: 2.5, z: 109.31942}
   m_LocalScale: {x: 10, y: 5, z: 5}
   m_Children:
   - {fileID: 4930518214379742}
   m_Father: {fileID: 0}
   m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: -180, z: 0}
 --- !u!4 &4930518214379742
 Transform:
   m_ObjectHideFlags: 1
@@ -87,7 +87,7 @@ MeshRenderer:
   m_LightProbeUsage: 1
   m_ReflectionProbeUsage: 1
   m_Materials:
-  - {fileID: 10303, guid: 0000000000000000f000000000000000, type: 0}
+  - {fileID: 2100000, guid: 3db83815f1a1de240ad3202375157882, type: 2}
   m_StaticBatchInfo:
     firstSubMesh: 0
     subMeshCount: 0
@@ -137,7 +137,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: cab2112ce93a3964d90c9d6013cfc8f0, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  pooledObject: {fileID: 1851471662954526, guid: 82ca7c422514696418f02b1e065e32e0,
+  pooledObject: {fileID: 1952406144847552, guid: 2e36be0614d476c4caf508f48d2d4cca,
     type: 2}
   pooledAmount: 10
   willGrow: 1
@@ -152,5 +152,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 3b853d60141e99f469476b099077db84, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  spawningTime: 2
-  wayPoints: []
+  spawningTime: 5
+  spawningPosition: {x: 0, y: 0, z: -10}
+  path: {fileID: 0}
