+++ Assets/Textures/Car/Spawner.mat
